replace CLI flags to a map in osvscanner library.
