+++ pkg/osvscanner/vulnerability_result.go
@@ -32,7 +32,6 @@ func buildVulnerabilityResults(
 	if len(licensesAllowlist) == 1 && licensesAllowlist[0] == "" {
 		licensesAllowlist = []string{}
 	}
-
 	for i, rawPkg := range packages {
 		includePackage := allPackages
 		var pkg models.PackageVulns
